+++ pkg/apis/options/sessions.go
@@ -5,6 +5,7 @@ type SessionOptions struct {
 	Type   string             `flag:"session-store-type" cfg:"session_store_type"`
 	Cookie CookieStoreOptions `cfg:",squash"`
 	Redis  RedisStoreOptions  `cfg:",squash"`
+	JWT    JWTStoreOptions    `cfg:",squash"`
 }
 
 // CookieSessionStoreType is used to indicate the CookieSessionStore should be
@@ -14,6 +15,10 @@ var CookieSessionStoreType = "cookie"
 // RedisSessionStoreType is used to indicate the RedisSessionStore should be
 // used for storing sessions.
 var RedisSessionStoreType = "redis"
+
+// JWTSessionStoreType is used to indicate the JWTSessionStore should be
+// used for storing sessions.
+var JWTSessionStoreType = "jwt"
 
 // CookieStoreOptions contains configuration options for the CookieSessionStore.
 type CookieStoreOptions struct {
@@ -32,6 +37,12 @@ type RedisStoreOptions struct {
 	ClusterConnectionURLs  []string `flag:"redis-cluster-connection-urls" cfg:"redis_cluster_connection_urls"`
 	CAPath                 string   `flag:"redis-ca-path" cfg:"redis_ca_path"`
 	InsecureSkipTLSVerify  bool     `flag:"redis-insecure-skip-tls-verify" cfg:"redis_insecure_skip_tls_verify"`
+}
+
+// JWTStoreOptions contains configuration options for the JWTSessionStore.
+type JWTStoreOptions struct {
+	JWTKey     string `flag:"jwt-session-key" cfg:"jwt_session_key"`
+	JWTKeyFile string `flag:"jwt-session-key-file" cfg:"jwt_session_key_file"`
 }
 
 func sessionOptionsDefaults() SessionOptions {