+++ pkg/apis/sessions/session_state.go
@@ -28,6 +28,7 @@ type SessionState struct {
 	User              string   `msgpack:"u,omitempty"`
 	Groups            []string `msgpack:"g,omitempty"`
 	PreferredUsername string   `msgpack:"pu,omitempty"`
+	Tenant            string   `msgpack:"t,omitempty"`
 }
 
 // IsExpired checks whether the session has expired
@@ -49,6 +50,9 @@ func (s *SessionState) Age() time.Duration {
 // String constructs a summary of the session state
 func (s *SessionState) String() string {
 	o := fmt.Sprintf("Session{email:%s user:%s PreferredUsername:%s", s.Email, s.User, s.PreferredUsername)
+	if s.Tenant != "" {
+		o += fmt.Sprintf(" tenant:%s", s.Tenant)
+	}
 	if s.AccessToken != "" {
 		o += " token:true"
 	}