+++ providers/providers.go
@@ -47,6 +47,8 @@ func New(provider string, p *ProviderData) Provider {
 		return NewDigitalOceanProvider(p)
 	case "google":
 		return NewGoogleProvider(p)
+	case "sis":
+		return NewSISProvider(p)
 	default:
 		return nil
 	}